fix: ack and write should be closely coupled
