+++ rust/numaflow-core/src/pipeline/isb/jetstream/writer.rs
@@ -4,7 +4,10 @@ use async_nats::jetstream::context::PublishAckFuture;
 use async_nats::jetstream::publish::PublishAck;
 use async_nats::jetstream::Context;
 use bytes::Bytes;
+use futures::channel::oneshot::Sender;
 use std::time::Duration;
+use tokio::sync::mpsc::Receiver;
+use tokio::sync::{mpsc, oneshot};
 use tokio::time::sleep;
 use tracing::error;
 
@@ -12,43 +15,75 @@ use tracing::error;
 /// Writes to JetStream ISB. Exposes both sync and async methods to write messages.
 pub(super) struct JetstreamWriter {
     js_ctx: Context,
+    paf_resolver_tx: mpsc::Sender<PublishResult>,
 }
 
 impl JetstreamWriter {
     pub(super) fn new(js_ctx: Context) -> Self {
-        Self { js_ctx }
+        let (paf_resolver_tx, paf_resolver_rx) = mpsc::channel::<PublishResult>(500);
+
+        let this = Self {
+            js_ctx,
+            paf_resolver_tx,
+        };
+
+        let mut resolver_actor = PafResolverActor::new(this.clone(), paf_resolver_rx);
+
+        tokio::spawn(async move {
+            resolver_actor.run().await;
+        });
+
+        this
     }
 
     /// Writes the message to the JetStream ISB and returns a future which can be
     /// awaited to get the PublishAck. It will do infinite retries until the message
     /// gets published successfully. If it returns an error it means it is fatal
     /// error
-    pub(super) async fn async_write(
+    pub(super) async fn write(
         &self,
         stream: &'static str,
         msg: Message,
-    ) -> Result<PublishAckFuture> {
+        success: oneshot::Sender<Result<()>>,
+    ) {
         let js_ctx = self.js_ctx.clone();
+        // this is for passing over for paf to do sync_write
+        let message = msg.clone();
+
         let payload = Bytes::from(
             msg.to_bytes()
                 .expect("message serialization should not fail"),
         );
-        // TODO: expose a way to exit the retry loop during shutdown
-        loop {
+        // FIXME: expose a way to exit the retry loop during shutdown
+        let paf = loop {
             match js_ctx.publish(stream, payload.clone()).await {
-                Ok(paf) => return Ok(paf),
+                Ok(paf) => {
+                    break paf;
+                }
                 Err(e) => {
-                    error!("publishing failed, retrying: {}", e);
+                    error!(?e, "publishing failed, retrying");
                     sleep(Duration::from_millis(10)).await;
                 }
             }
-        }
+        };
+
+        let result = PublishResult {
+            paf,
+            stream,
+            message,
+            callee_tx: success,
+        };
+
+        self.paf_resolver_tx
+            .send(result)
+            .await
+            .expect("send should not fail");
     }
 
     /// Writes the message to the JetStream ISB and returns the PublishAck. It will do
     /// infinite retries until the message gets published successfully. If it returns
-    /// an error it means it is fatal error
-    pub(super) async fn sync_write(
+    /// an error it means it is fatal non-retryable error.
+    pub(super) async fn blocking_write(
         &self,
         stream: &'static str,
         msg: Message,
@@ -74,6 +109,52 @@ impl JetstreamWriter {
                     sleep(Duration::from_millis(10)).await;
                 }
             }
+        }
+    }
+}
+
+/// PublishResult is the result of the write operation.
+/// It contains the PublishAckFuture which can be awaited to get the PublishAck.
+#[derive(Debug)]
+pub(super) struct PublishResult {
+    paf: PublishAckFuture,
+    stream: &'static str,
+    message: Message,
+    callee_tx: oneshot::Sender<Result<()>>,
+}
+
+struct PafResolverActor {
+    js_writer: JetstreamWriter,
+    receiver: Receiver<PublishResult>,
+}
+
+impl PafResolverActor {
+    fn new(js_writer: JetstreamWriter, receiver: Receiver<PublishResult>) -> Self {
+        PafResolverActor {
+            js_writer,
+            receiver,
+        }
+    }
+    async fn handle_result(&mut self, result: PublishResult) {
+        match result.paf.await {
+            Ok(ack) => result.callee_tx.send(Ok(())).unwrap(),
+            Err(e) => {
+                error!("Failed to resolve the future, trying sync write");
+                match self
+                    .js_writer
+                    .blocking_write(result.stream, result.message.clone())
+                    .await
+                {
+                    Ok(_) => result.callee_tx.send(Ok(())).unwrap(),
+                    Err(e) => result.callee_tx.send(Err(e)).unwrap(),
+                }
+            }
+        }
+    }
+
+    async fn run(&mut self) {
+        while let Some(result) = self.receiver.recv().await {
+            self.handle_result(result).await;
         }
     }
 }
@@ -118,12 +199,13 @@ mod tests {
             headers: HashMap::new(),
         };
 
-        let result = writer.async_write(stream_name, message.clone()).await;
-        assert!(result.is_ok());
+        let (success_tx, success_rx) = oneshot::channel::<Result<()>>();
+        writer.write(stream_name, message.clone(), success_tx).await;
+        assert!(success_rx.await.is_ok());
 
-        let publish_ack_future = result.unwrap();
-        let publish_ack = publish_ack_future.await;
-        assert!(publish_ack.is_ok());
+        // let publish_ack_future = result.unwrap();
+        // let publish_ack = publish_ack_future.await;
+        // assert!(publish_ack.is_ok());
 
         context.delete_stream(stream_name).await.unwrap();
     }
@@ -159,7 +241,7 @@ mod tests {
             headers: HashMap::new(),
         };
 
-        let result = writer.sync_write(stream_name, message.clone()).await;
+        let result = writer.blocking_write(stream_name, message.clone()).await;
         assert!(result.is_ok());
 
         let publish_ack = result.unwrap();
